tests: Work around mac80211_hwsim survey limitations
ap_multi_bss_acs test case was failing if it was executed immediately
after a test case that used 5 GHz band since the current mac80211_hwsim
channel survey implementation is very limited and reports only a single
channel. For this test case, that channel has to be from the 2.4 GHz
band.

Signed-off-by: Jouni Malinen <j@w1.fi>