nl80211: Fix non-hostapd interface addition to not call add_ifidx()
Commit b36935be1a14341771b0fd5491808c3f6fdcb603 ('nl80211: Fix EAPOL
frames not being delivered') and commit
147848ec4d26613d5a117d4b35dbc7ff98dd65d1 ('nl80211: Do not add all
virtual interfaces to drv->if_indices') were not fully in sync and it
was possible for some non-hostapd use cases to end up adding undesired
ifindexes into the list of interfaces from which events and EAPOL frames
are processed on the parent interface. This could result, e.g., in P2P
Device management interface on getting unexpected events, including
RTM_NEWLINK event that could end up getting interpreted as an
indication of the interface being down and unavailable.

Make both add_ifidx() calls use the same criteria for adding interfaces
to the local list. This is not really a complete solution, but it is
good enough for now to fix the most visible side effects of this issue.

Signed-off-by: Jithu Jance <jithu@broadcom.com>